Improvements to the configure script
The configure script will now check for mysql and libev if any of them
are required. A short summary of what will be compiled and installed
will be displayed to the user once the configure script finishes.

Also adapted mb-clientd, mb-masterd and mb-slaved to use installed
versions of libev instead, since the bundled version was removed.